feat(orchestrate-core): move WorkflowState into the core on the pure event (#109)
Slice 2 of the Event-ABI round (noetl/ai-meta#109): `state.rs` (WorkflowState +
the from_events/apply_event drive-state machine) moves into noetl-orchestrate-core
on `core::event::Event`, so the drive state reconstruction compiles to wasm32.

Wiring:
- `state.rs` uses `core::event::Event`; `event.created_at` → `event.timestamp`.
- The server converts `db::Event → core::event::Event` at the drive boundary:
  orchestrator's `evaluate` and the four `from_events`/`apply_event` sites in
  events.rs map the bounded slice through `From<&db::Event>` (cheap — refs-in-state
  keeps payloads small).  `orchestrator` keeps its `db::Event` signature (its big
  fixture set is untouched); its 11 test apply_event calls convert at the call.
- wasm portability: `Instant` self-timing in `from_events` is `#[cfg(not(wasm32))]`
  (the plug-in host times the invoke); `tracing` added to the core (wasm-safe);
  `catalog_id` added to `core::event::Event` (the read-set seeds WorkflowState
  with it), `#[serde(default)]` so the wire tolerates its absence.
- the one state+orchestrator integration test moved to orchestrator.rs's test
  module (it needs the server `WorkflowOrchestrator`); state's `pub(crate)`
  helpers the server reaches are now `pub`.

587 server + 86 core tests green; core compiles to wasm32-unknown-unknown (no
WASI); clippy clean.  Pure refactor.  Last slice: move `orchestrator`/`evaluate`.

Refs noetl/ai-meta#109
