+++ orchestrate-core/Cargo.toml
@@ -16,6 +16,7 @@ serde_json = "1.0"
 base64 = "0.22"
 chrono = { version = "0.4", default-features = false, features = ["serde", "std"] }
 thiserror = "2.0"
+tracing = "0.1"
 
 [dev-dependencies]
 serde_yaml = "0.9"
